requirements.txt
# Core ML/AI
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
accelerate>=0.24.0
peft>=0.7.0

# Vector Stores
faiss-cpu>=1.7.4
qdrant-client>=1.7.0

# LangChain
langchain>=0.1.0
langchain-community>=0.0.10

# Training (optional for HF Spaces)
# trl>=0.7.0
# datasets>=2.14.0

# UI
streamlit>=1.29.0

# Utils
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
